backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.6.15
click==8.1.8
cloudinary==1.44.1
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
motor==3.6.1
passlib==1.7.4
pkg_resources==0.0.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.33.0
